+++ Assets/Demo/PalmSwipeDemo.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using FluentMotion.extensions;
 using FluentMotion.hand;
 using FluentMotion.helpers;
 using Leap;
@@ -6,26 +7,28 @@ using Leap.Unity;
 using UniRx;
 using UnityEngine;
 using UnityEngine.Serialization;
+using Valve.VR.InteractionSystem;
 
 namespace Demo
 {
     public class PalmSwipeDemo : MonoBehaviour
     {
-        private GameObject _cube;
-        
-        [FormerlySerializedAs("Hand")] 
-        public ReactiveHand hand;
+        [FormerlySerializedAs("Cube")] public GameObject cubes;
+
+        [FormerlySerializedAs("Hand")] public ReactiveHand hand;
 
         public void Start()
         {
-            _cube = GameObject.CreatePrimitive(PrimitiveType.Cube);
-            _cube.transform.Translate(Vector3.back * 5);
-
             hand.AsObservable()
-                .Where(h => h.PalmVelocity.AngleTo(Vector.Right) < Math.PI / 6)
-                .Where(h => h.PalmVelocity.Magnitude > 0.01)
-                .Select(h => Rotation.Right(h.PalmVelocity.Magnitude.Map(0.01f, 1.0f, 0.1f, 2f)))
-                .Subscribe(rotation => _cube.transform.Rotate(rotation.eulerAngles));
+                .IsMovingDown(Player.instance)
+                .Select(h => Rotation.Right(h.PalmVelocity.ToVector3().magnitude.Map(0.1f, 2, 0.1f, 2f)))
+                .Subscribe(rotation =>
+                           {
+                               foreach (Transform child in cubes.transform)
+                               {
+                                   child.Rotate(rotation.eulerAngles);
+                               }
+                           });
         }
     }
 }