+++ Assets/FluentMotion/helpers/Mixins.cs
@@ -1,10 +1,11 @@
 using System;
+using Leap.Unity;
+using UnityEngine;
 
 namespace FluentMotion.helpers
 {
     public static class Mixins
     {
-        public static float Degrees(this float @this) => (float) (@this * 180.0f  / Math.PI);
-        public static float Radians(this float @this) => (float) (@this * Math.PI / 180.0f);
+        public static bool Equals(this float @this, float other, float epsilon = float.Epsilon) => Math.Abs(@this - other) < epsilon;
     }
 }